% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gap.R
\name{plot_organ_gap}
\alias{plot_organ_gap}
\title{Plot Organ Age Gap Distribution}
\usage{
plot_organ_gap(data)
}
\arguments{
\item{data}{Data frame with HeartAgeGap_ML, KidneyAgeGap_ML,
MetabAgeGap_ML.}
}
\value{
A ggplot object.
}
\description{
Density plot of organ age gaps for Heart, Kidney, and Metab.
}
