% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organ_age.R
\name{predict_organ_age}
\alias{predict_organ_age}
\title{Predict Organ-Specific Biological Age (Path A)}
\usage{
predict_organ_age(data, organ = c("Heart", "Kidney", "MetabInf"))
}
\arguments{
\item{data}{Data frame with clinical biomarkers.}

\item{organ}{One of 'Heart', 'Kidney', 'MetabInf'.}
}
\value{
Data frame with new column '\if{html}{\out{<organ>}}Age_ML'.
For MetabInf, the column is named 'MetabAge_ML'.
}
\description{
Predict Organ-Specific Biological Age (Path A)
}
